consolidate trainer service
